pyproject.toml
[tool.poetry]
name = "top-api"
version = "0.1.0"
description = ""
authors = ["victro-nuzhniy <nuzhniyva@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
coreapi = "^2.3.3"
djangorestframework-simplejwt = "^5.3.0"
psycopg2-binary = "^2.9.7"
pyyaml = "^6.0.1"
celery = "^5.3.4"
redis = "^5.0.0"
drf-yasg = "^1.21.7"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
flake8-docstrings = "^1.7.0"
isort = "^5.12.0"
pre-commit = "^3.4.0"


[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.4.1"
pytest-mock = "^3.11.1"
pytest-sugar = "^0.9.7"
pytest-cov = "^4.1.0"
pytest-randomly = "^3.15.0"
pytest-clarity = "^1.0.1"
faker = "^19.3.1"
pytest-django = "^4.5.2"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  | dist
  | .pytest_cache
  | .tox
  | docs
  | docs_src
  | .venv
  | .mypy_cache
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = []

[tool.pytest.ini_options]
# === Pytest ===
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = [".", "top_api"]
norecursedirs = ["env", ".env", "migrations", "venv"]
testpaths = ["tests",]
console_output_style = "progress"
python_files = "test_*"
python_classes = "Test*"
python_functions = "test_*"

# === Pytest ===